Add java docs for TokenRepository and dependent classes
